+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 5.0.18 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix release
 
 
 5.0.17 (2021-02-04)