+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-6.0.2 (unreleased)
+6.0.2 (2021-05-05)
 ------------------
 
 - Allow to delete multiattachment fields